Modified data extraction to discard frames labeld as 'q'
The 'q' label indicates a glottal stop. Most papers I've seen discard the frames that are
labeled like this so as to simplify the modeling process. And in fact, most of the times
that 'q' would show up in English, it is really just an allophone of another sound, so
we should be able to model English to some approximate degree without it.