Fixed evaluateAccuracy function
The evaluateAccuracy function was not accessing the data of the TrackedReal objects
that are output from the model correctly. I found out there is a data() function to
extract the values from Flux's Tracked objects, which made the routine actually
run correctly. Also, at the outset, I was using indmax() to find the class, but
since there was only 1 element, it was always 1, which produced the 100% accuracy.
This is fixed now as well.